# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "园智汇 · AI 产业园区智慧运营管理平台",
  "description": "SmartPark AI Operations Platform\n\n面向产业园区集团的多租户智慧运营平台，包含数据驾驶舱、空间资产、企业服务、招商运营、合同财务、项目管理、物业运维、安全管理、审批中心、AI Agent（1 个主 Agent + 11 个子 Agent）、数据中心、报表中心、日志审计与系统管理。\n\n**数据说明**：演示环境，全部数据为演示数据（is_demo=True），由数据库实时聚合计算。",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/leasing/channels": {
   "get": {
    "tags": [
     "招商租赁"
    ],
    "summary": "List Channels",
    "operationId": "list_channels_api_leasing_channels_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Channels Api Leasing Channels Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/funnel": {
   "get": {
    "tags": [
     "招商租赁"
    ],
    "summary": "Funnel",
    "description": "招商漏斗：各阶段转化率 + 各园区对比。",
    "operationId": "funnel_api_leasing_funnel_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Funnel Api Leasing Funnel Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/leads": {
   "get": {
    "tags": [
     "招商租赁"
    ],
    "summary": "List Leads",
    "operationId": "list_leads_api_leasing_leads_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "stage",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Stage"
      }
     },
     {
      "name": "channel_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Channel Id"
      }
     },
     {
      "name": "owner",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Owner"
      }
     },
     {
      "name": "keyword",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Keyword"
      }
     },
     {
      "name": "stagnant_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Stagnant Only"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Leads Api Leasing Leads Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/leads/{lead_id}": {
   "get": {
    "tags": [
     "招商租赁"
    ],
    "summary": "Lead Detail",
    "operationId": "lead_detail_api_leasing_leads__lead_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lead_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Lead Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Lead Detail Api Leasing Leads  Lead Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/leads/{lead_id}/followup": {
   "post": {
    "tags": [
     "招商租赁"
    ],
    "summary": "Add Followup",
    "operationId": "add_followup_api_leasing_leads__lead_id__followup_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lead_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Lead Id"
      }
     },
     {
      "name": "content",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Content"
      }
     },
     {
      "name": "followup_type",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "电话跟进",
       "title": "Followup Type"
      }
     },
     {
      "name": "result",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Result"
      }
     },
     {
      "name": "next_action",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Next Action"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add Followup Api Leasing Leads  Lead Id  Followup Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/leads/{lead_id}/stage": {
   "post": {
    "tags": [
     "招商租赁"
    ],
    "summary": "Move Stage",
    "description": "推进线索阶段（含审计与合法性校验）。",
    "operationId": "move_stage_api_leasing_leads__lead_id__stage_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lead_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Lead Id"
      }
     },
     {
      "name": "stage",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Stage"
      }
     },
     {
      "name": "lost_reason",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Lost Reason"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Move Stage Api Leasing Leads  Lead Id  Stage Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/activities": {
   "get": {
    "tags": [
     "招商租赁"
    ],
    "summary": "List Activities",
    "operationId": "list_activities_api_leasing_activities_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Activities Api Leasing Activities Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/leasing/stagnant": {
   "get": {
    "tags": [
     "招商租赁"
    ],
    "summary": "Stagnant Leads",
    "description": "停滞线索分析（供 AI 与看板使用）。",
    "operationId": "stagnant_leads_api_leasing_stagnant_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 30,
       "title": "Days"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Stagnant Leads Api Leasing Stagnant Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/login"
     }
    }
   }
  }
 }
}